Status order changed

+++ pojo/src/main/java/com/stanzaliving/core/leadership/dto/ListingDto.java
@@ -64,7 +64,7 @@ public class ListingDto {
 
 	private Long rentPerBed;
 	
-	private Long rent;
+	private Integer rent;
 
 	private Date llHandoverDate;
